Update C2Cnew.py
Added the capability to install Nginx.
Added the ability to upload a website.
Added the ability to obtain an SSL certificate.
Fixed the issue of changing the UDP port, which was previously possible only once. Now, you can change the port an infinite number of times.
Fixed the issue with the server tunneling, which was previously only possible once. Now, you can tunnel as many times as you want to as many servers as you need.
Overall improvements to the program.